backend/requirements.txt
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
backports-datetime-fromisoformat==2.0.3
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
exceptiongroup==1.3.1
Flask==3.1.2
flask-cors==6.0.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.1.2
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.45
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
